Moved Notifications and Response structs to be a subpackage of datahandling
