+++ modules/datahandling/userrequests.go
@@ -1,8 +1,8 @@
 package datahandling
 
 import (
+	"github.com/CodeCollaborate/Server/modules/datahandling/messages"
 	"github.com/CodeCollaborate/Server/modules/dbfs"
-	"github.com/CodeCollaborate/Server/modules/messages"
 	"github.com/CodeCollaborate/Server/modules/rabbitmq"
 	"golang.org/x/crypto/bcrypt"
 )
